requirements.txt
dependency-injector
flask
Flask-Mail
PyJWT
minio
requests
SQLAlchemy-serializer 
Flask-SQLAlchemy